+++ frontend/frontend/src/Components/Home.js
@@ -1,19 +1,15 @@
-import React, { Component } from 'react';
+import React from 'react';
 import Home from './home/Home'
 import NavHome from './home/NavHome'
 import './styles/nav.css'
 import './styles/home.css'
 
-
-class home extends Component {
-    render() {
-        return (
-            <>
-                <NavHome />
-                <Home />
-            </>
-        )
-    }
-
+function home() {
+    return (
+        <>
+            <NavHome />
+            <Home />
+        </>
+    )
 }
 export default home;
