+++ frontend/frontend/src/Components/Products.js
@@ -1,19 +1,16 @@
-import React, { Component } from 'react';
+import React from 'react';
 import Products from './products/Products'
 import NavMenu from './Layout/NavMenu'
 import './styles/nav.css'
 import './styles/products.css'
 
 
-class products extends Component {
-    render() {
-        return (
-            <>
-                <NavMenu />
-                <Products />
-            </>
-        )
-    }
-
+function products() {
+    return (
+        <>
+            <NavMenu />
+            <Products />
+        </>
+    )
 }
 export default products;
