+++ frontend/frontend/src/Components/company.js
@@ -1,19 +1,15 @@
-import React, { Component } from 'react';
+import React from 'react';
 import Company from './company/ChooseCompany'
 import NavMenu from './Layout/NavMenu'
 import './styles/nav.css'
 import './styles/chooseCompany.css'
 
-
-class company extends Component {
-    render() {
-        return (
-            <>
-                <NavMenu />
-                <Company />
-            </>
-        )
-    }
-
+function company() {
+    return (
+        <>
+            <NavMenu />
+            <Company />
+        </>
+    )
 }
 export default company;
