+++ frontend/frontend/src/Components/profile.js
@@ -1,19 +1,15 @@
-import React, { Component } from 'react';
+import React from 'react';
 import Profile from './profile/UpdateProfile'
 import './styles/nav.css'
 import './styles/home.css'
 import NavMenu from './Layout/NavMenu';
 
-
-class profile extends Component {
-    render() {
-        return (
-            <>
-                <NavMenu />
-                <Profile />
-            </>
-        )
-    }
-
+function profile() {
+    return (
+        <>
+            <NavMenu />
+            <Profile />
+        </>
+    )
 }
 export default profile;
