+++ frontend/frontend/src/Components/signup.js
@@ -1,19 +1,15 @@
-import React, { Component } from 'react';
+import React from 'react';
 import Register from './signup/Register'
-import NavLogin from './login/NavLogin'
+import NavLogin from './Layout/NavLogin'
 import './styles/nav.css'
 import './styles/login.css'
 
-
-class Signup extends Component {
-    render() {
-        return (
-            <>
-                <NavLogin />
-                <Register />
-            </>
-        )
-    }
-
+function Signup() {
+    return (
+        <>
+            <NavLogin />
+            <Register />
+        </>
+    )
 }
 export default Signup;
