--- frontend/frontend/src/Reducers/rootReducer.js
@@ -1,9 +1,0 @@
-const initState = {
-    post: []
-}
-
-const rootReducer = (state = initState, action) => {
-    return state
-}
-
-export default rootReducer
