+++ frontend/frontend/src/index.js
@@ -4,7 +4,7 @@ import App from './App';
 import '../node_modules/font-awesome/css/font-awesome.min.css';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
-import rootReducer from './Reducers/rootReducer'
+import rootReducer from './store/Reducers/rootReducer'
 
 const store = createStore(rootReducer)
