+++ frontend/frontend/src/store/Reducers/authReducer.js
@@ -1,0 +1,9 @@
+const initState = {
+
+}
+
+const authReducer = (state = initState, action) => {
+    return state
+}
+
+export default authReducer
