+++ frontend/frontend/src/store/Reducers/projectReducer.js
@@ -1,0 +1,9 @@
+const initState = {
+
+}
+
+const projectReducer = (state = initState, action) => {
+    return state
+}
+
+export default projectReducer
