+++ frontend/frontend/src/store/Reducers/rootReducer.js
@@ -1,0 +1,11 @@
+import authReducer from './authReducer'
+import projectReducer from './projectReducer'
+import { combineReducers } from 'redux'
+
+const rootReducer = combineReducers({
+    auth: authReducer,
+    project: projectReducer
+})
+
+
+export default rootReducer
